+++ pages/side-panel/src/pages/chat-room.tsx
