+++ pages/side-panel/src/side-panel.css
