+++ pages/side-panel/src/side-panel.tsx
@@ -1,7 +1,7 @@
-import '@src/SidePanel.css';
+import '@src/side-panel.css';
 import { withErrorBoundary, withSuspense } from '@extension/shared';
 import { SignedIn, SignedOut } from '@clerk/chrome-extension';
-import Chatroom from './pages/chatroom';
+import Chatroom from './pages/chat-room';
 import Signin from './pages/signin';
 import { SelectedTabsProvider } from './hooks/use-selected-tabs';
 const SidePanel = () => {
